refactor(settings): extract connector settings ownership
